fix: declutter mode picker and use full month names in version stamp

+++ src/main.js
@@ -11,7 +11,7 @@ function renderBuildInfo() {
   if (el && typeof __TEPUQ_BUILD_TIME__ !== 'undefined') {
     const d = new Date(__TEPUQ_BUILD_TIME__);
     const fmt = d.toLocaleString('id-ID', {
-      day: '2-digit', month: 'short', year: 'numeric',
+      day: '2-digit', month: 'long', year: 'numeric',
       hour: '2-digit', minute: '2-digit',
     });
     el.textContent = `Versi ${fmt}`;
